feat: add --export-to=PATH to save bulk exports
This offers a way to save the results of a bulk export locally,
rather that deleting them when done with them.

Useful for debugging or archiving.
