+++ cumulus/loaders/__init__.py
@@ -1,5 +1,5 @@
 """Public API for loaders"""
 
 from .base import Loader
-from .fhir.fhir_ndjson import FhirNdjsonLoader
+from .fhir.ndjson_loader import FhirNdjsonLoader
 from .i2b2.loader import I2b2Loader
