+++ cumulus/loaders/base.py
@@ -1,10 +1,22 @@
 """Base abstract loader"""
 
 import abc
-import tempfile
-from typing import List
+from typing import List, Protocol
 
 from cumulus.store import Root
+
+
+# Loaders often want the ability to provide callers with a TemporaryDirectory -- a folder that will be deleted as soon
+# as the caller is done using it. However, some code flows may instead require returning a folder that already exists.
+# So for that use case, we define a Directory protocol for typing purposes, which looks like a TemporaryDirectory.
+# And then a RealDirectory class that does not delete its folder.
+class Directory(Protocol):
+    name: str
+
+
+class RealDirectory:
+    def __init__(self, path: str):
+        self.name = path
 
 
 class Loader(abc.ABC):
@@ -24,7 +36,7 @@ class Loader(abc.ABC):
         self.root = root
 
     @abc.abstractmethod
-    async def load_all(self, resources: List[str]) -> tempfile.TemporaryDirectory:
+    async def load_all(self, resources: List[str]) -> Directory:
         """
         Loads the listed remote resources and places them into a local folder as FHIR ndjson
